Remove Fight file
